communion/stt/requirements.txt
openai-whisper
sounddevice
numpy
requests
